refactor: move task logic into tasks folder
